AutoChem

Documentation for AutoChem.

AutoChem.BimolecularReaction

AutoChem.TrimolecularReaction

AutoChem.parse_bimol_d

AutoChem.parse_photolysis_d

AutoChem.parse_trimol_d

AutoChem.read_bimol

AutoChem.read_photolysis

AutoChem.read_trimol

AutoChem.BimolecularReaction — Method

(rxn::BimolecularReaction)(T,P,M)

Given a reaction rxn of type BimolecularReaction, compute the reaction rate coefficient as a function of

T: temperature in Kelvin

P: pressure in mbar

M: the total particle number density

AutoChem.TrimolecularReaction — Method

(rxn::TrimolecularReaction)(T,P,M)

Given a reaction rxn of type TrimolecularReaction, compute the reaction rate coefficient as a function of

T: temperature in Kelvin

P: pressure in mbar

M: the total particle number density

AutoChem.parse_bimol_d — Method

parse_bimol_d(path)

Parse an AutoChem bimolecular reaction database, returning a vector of BimolecularReaction objects and a second vector containing the indices of any reactions that failed to parse.

AutoChem.parse_photolysis_d — Method

parse_photolysis_d(path)

Parse an AutoChem photolysis reaction database, returning a vector of PhotolysisReaction objects and a second vector containing the indices of any reactions that failed to parse.

AutoChem.parse_trimol_d — Method

parse_trimol_d(path)

Parse an AutoChem trimolecular reaction database, returning a vector of TrimolecularReaction objects and a second vector containing the indices of any reactions that failed to parse.

AutoChem.read_bimol — Method

read_bimol(path)

Parse a bimolecular reaction database (in JSON format) returning a vector of BimolecularReaction objects.

AutoChem.read_photolysis — Method

read_photolysis(path)

Parse a photolysis reaction database (in JSON format) returning a vector of PhotolysisReaction objects.

AutoChem.read_trimol — Method

read_trimol(path)

Parse a trimolecular reaction database (in JSON format) returning a vector of TrimolecularReaction objects.